3/pyproject.toml
[tool.poetry]
name = "ollama-server"
version = "0.1.0"
description = "Uniwersalny serwer dla modeli Ollama z interfejsem API"
authors = ["Tom Sapletta <info@softreck.dev>"]
readme = "README.md"
license = "Apache"
repository = "https://github.com/username/ollama-server"
keywords = ["ollama", "llm", "ai", "api", "server"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
packages = [
    {include = "ollama_server"}
]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
click = "^8.1.7"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
ollama-server = "ollama_server.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]